Make it possible to choose the renderer at runtime, with a constant.
