+++ tests/playground.html
@@ -71,6 +71,7 @@
 var workspace = null;
 
 function start() {
+  Blockly.renderMode = 'compatibility';
   setBackgroundColour();
 
   // Parse the URL arguments.
